25-07

+++ Application/DTOs/RequestDTOs/IO/PressureItemDTO.cs
@@ -3,7 +3,7 @@
     public class PressureItemDTO
     {
         public string Port { get; set; }
-        public int Pressure { get; set; }
+        public double Pressure { get; set; }
         public DateTime CreateDate { get; set; } = DateTime.Now;
     }
 }
